+++ lib/services/colors.dart
@@ -4,5 +4,5 @@ abstract class CustomColors {
   static Color maroon = const Color(0xFF8B2635);
   static Color yellow = const Color(0xFFFFF5AE);
   static Color blue = const Color(0xFFD2C2FF);
-  static Color gray = const Color(0xFF9E9E9E);
+  static Color lightGray = const Color(0xFF9E9E9E);
 }
